backend/requirements.txt
asgiref==3.11.1
Django==5.0.4
djangorestframework==3.16.1
django-cors-headers==4.9.0
pdfplumber==0.11.9
sqlparse==0.5.5
tzdata==2025.3
spacy==3.7.4
djangorestframework-simplejwt==5.3.1
easyocr==1.7.1
opencv-python-headless==4.11.0.86
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
google-auth==2.39.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
whitenoise==6.8.2